Immutability for blake2 + test_freeze module
