+++ Lib/test/test_freeze/__init__.py
@@ -1,0 +1,27 @@
+import os
+from test.support import load_package_tests
+import unittest
+
+
+def load_tests(*args):
+    return load_package_tests(os.path.dirname(__file__), *args)
+
+
+class BaseObjectTest(unittest.TestCase):
+    def __init__(self, *args, obj=None, **kwargs):
+        unittest.TestCase.__init__(self, *args, **kwargs)
+        self.obj = obj
+
+    def setUp(self):
+        freeze(self.obj)
+
+    def test_immutable(self):
+        self.assertTrue(isimmutable(self.obj))
+
+    def test_add_attribute(self):
+        freeze(self.obj)
+        with self.assertRaises(NotWriteableError):
+            self.obj.new_attribute = 'value'
+
+    def test_type_immutable(self):
+        self.assertTrue(isimmutable(type(self.obj)))
