+++ Lib/test/test_freeze/test_hashlib.py
@@ -1,0 +1,18 @@
+from hashlib import blake2b, blake2s
+import unittest
+
+
+class TestHashlib(unittest.TestCase):
+    def test_blake2b(self):
+        h = blake2b(digest_size=32)
+        h.update(b'Hello world')
+        freeze(h)
+        with self.assertRaises(NotWriteableError):
+            h.update(b'!')
+
+    def test_blake2s(self):
+        h = blake2s(digest_size=32)
+        h.update(b'Hello world')
+        freeze(h)
+        with self.assertRaises(NotWriteableError):
+            h.update(b'!')
